Moved converted pages from Matthew-Angular. (ERRORS lost-and-found)
-ERRORS
THIS COMMIT THROWS ERRORS FROM the lost-and-found files. The errors come from lines 51-63 in lost-and-found.component.html. These lines have been commented out for now.

Moved the pages converted to angular from Matthew-Angular to this branch, and updated the routing to reflect the new file system. Also, added ":standalone true" statements to component.ts files that gave errors from having imports but not being explicitly standalone.

-add-item page
Also, moved the items page info into the add-item page. The add-item page was a non-functional duplicate of the items page, but the name made more sense. So now the add-items page is the old items page.